refactor: improve typescript types
