+++ src/helpers/divider-helper.ts
@@ -1,3 +1,3 @@
-export function showSectionDivider(array: any, index: number): boolean {
+export function showSectionDivider<T>(array: T[], index: number): boolean {
   return array.length > 1 && index !== array.length - 1;
 }
